requirements.txt
Django==5.0.7
djangorestframework==3.15.2
requests==2.32.3
psycopg2-binary==2.9.9 
gunicorn==22.0.0 
